Fix unit test warnings and circular dependencies
- Upgraded `pytest` to 9.0.2 in `test_requirements.txt` to resolve compatibility warnings with Python 3.12 (`TestResult has no addDuration method`).
- Fixed `SAWarning` in `timesketch/lib/testlib.py` by removing redundant engine recreation in `ModelBaseTest`, which conflicted with the existing session.
- Moved `get_yaml_config` and `get_config_path` from `timesketch/lib/analyzers/interface.py` to `timesketch/lib/utils.py` to resolve circular import dependencies between `interface`, `ontology`, and `api/v1/utils`.
- Updated imports in `timesketch/lib/analyzers/interface.py` and `timesketch/lib/ontology.py`.
- Fixed `FutureWarning` in `timesketch/lib/analyzers/authentication/utils.py` by removing `inplace=True` from `fillna` and `sort_values` calls on Pandas DataFrames.
- Fixed `DeprecationWarning` in `timesketch/lib/llms/features/nl2q.py` by using `include_groups=False` in `groupby().apply()` and updating the callback to access the group name correctly.
