+++ test_requirements.txt
@@ -1,6 +1,6 @@
 Flask-Testing >= 0.6.2
 mock >= 2.0.0
-pytest==7.1.3
+pytest==9.0.2
 pbr >= 4.2.0
 beautifulsoup4 >= 4.8.2
 coverage >= 5.0.2
