+++ timesketch/lib/ontology.py
@@ -15,12 +15,12 @@
 
 import json
 
-from timesketch.lib.analyzers import interface
+from timesketch.lib import utils
 
 
 def ontology():
     """Return a dict with the ontology definitions."""
-    return interface.get_yaml_config("ontology.yaml")
+    return utils.get_yaml_config("ontology.yaml")
 
 
 class OntologyInterface:
