+++ timesketch/lib/utils.py
@@ -23,8 +23,10 @@ import random
 import smtplib
 import time
 import codecs
+import os
 from typing import List, Optional
 import pandas
+import yaml
 
 from dateutil import parser
 from flask import current_app
@@ -737,3 +739,52 @@ def send_email(subject: str, body: str, to_username: str, use_html: bool = False
         smtp.login(email_login_username, email_login_password)
     smtp.sendmail(msg["From"], [msg["To"]], msg.as_string())
     smtp.quit()
+
+
+def get_config_path(file_name):
+    """Returns a path to a configuration file.
+
+    Args:
+        file_name: String that defines the config file name.
+
+    Returns:
+        The path to the configuration file or None if the file cannot be found.
+    """
+    path = os.path.join(os.path.sep, "etc", "timesketch", file_name)
+    if os.path.isfile(path):
+        return path
+
+    path = os.path.join(os.path.dirname(__file__), "..", "..", "data", file_name)
+    path = os.path.abspath(path)
+    if os.path.isfile(path):
+        return path
+
+    return None
+
+
+def get_yaml_config(file_name: str):
+    """Return a dict parsed from a YAML file within the config directory.
+
+    Args:
+        file_name: String that defines the config file name.
+
+    Returns:
+        A dict with the parsed YAML content from the config file or
+        an empty dict if the file is not found or YAML was unable
+        to parse it.
+    """
+    path = get_config_path(file_name)
+    if not path:
+        return {}
+
+    with open(path, "r", encoding="utf-8") as fh:
+        try:
+            return yaml.safe_load(fh)
+        except yaml.parser.ParserError as exception:
+            # pylint: disable=logging-format-interpolation
+            logger.warning(
+                ("Unable to read in YAML config file, " "with error: {!s}").format(
+                    exception
+                )
+            )
+            return {}
